primitives: move VarInt into coding module
